+++ src/com/mwpb/Aoc19.java
@@ -1,22 +1,27 @@
 package com.mwpb;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 public class Aoc19 {
 	String intcodeString;
 	Map<Point, Boolean> cache;
-	
+
 	Aoc19(String intcodeString) {
 		this.intcodeString = intcodeString;
 		this.cache = new HashMap<>();
 	}
-	
+
 	boolean checkPoint(Point p) {
 		if (this.cache.containsKey(p)) {
 			return this.cache.get(p);
 		} else {
 			Intcode intcode = new Intcode(this.intcodeString);
-			intcode.input = new LinkedList<Integer>(List.of(p.x, p.y));
+			intcode.input = new LinkedList<Long>(
+					List.of((long) p.x, (long) p.y));
 			Optional<Long> output = intcode.run();
 			boolean isInBeam = output.get() == 1;
 
@@ -33,7 +38,7 @@ public class Aoc19 {
 			System.out.print("\n");
 		}
 	}
-	
+
 	int countIn50x50() {
 		int count = 0;
 		for (int i = 0; i < 50; i++) {
@@ -49,7 +54,8 @@ public class Aoc19 {
 	boolean checkBox(Point p) {
 		for (int i = 0; i < 100; i++) {
 			for (int j = 0; j < 100; j++) {
-				// System.out.println(String.format("%s: %b", new Point(i + p.x, j + p.y), this.checkPoint(new Point(i + p.x, j + p.y))));
+				// System.out.println(String.format("%s: %b", new Point(i + p.x,
+				// j + p.y), this.checkPoint(new Point(i + p.x, j + p.y))));
 				if (!this.checkPoint(new Point(i + p.x, j + p.y))) {
 					return false;
 				}
@@ -60,16 +66,17 @@ public class Aoc19 {
 
 	Optional<Point> checkRow(int y) {
 		int x = 0;
-		while(!this.checkPoint(new Point(x, y))) {
+		while (!this.checkPoint(new Point(x, y))) {
 			x++;
 		}
 		int first = x;
-		while(this.checkPoint(new Point(x, y))) {
+		while (this.checkPoint(new Point(x, y))) {
 			x++;
 		}
 		int last = x;
-		// System.out.println(String.format("%d - %d = %d", last, first, last - first));
-		for (int i = first ; i < last + 1 - 100; i++) {
+		// System.out.println(String.format("%d - %d = %d", last, first, last -
+		// first));
+		for (int i = first; i < last + 1 - 100; i++) {
 			Point p = new Point(i, y);
 			if (this.checkBox(p)) {
 				return Optional.of(p);
@@ -88,5 +95,5 @@ public class Aoc19 {
 		} while (pOption.isEmpty() && y < 10000);
 		return pOption.get();
 	}
-	
+
 }
